Refactor - change namespaces and folder layout in infrastructure project
