+++ src/SampleProject.API/Startup.cs
@@ -9,7 +9,7 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Configuration.UserSecrets;
 using Microsoft.Extensions.DependencyInjection;
 using SampleProject.API.Modules;
-using SampleProject.Infrastructure.Orders;
+using SampleProject.Infrastructure;
 
 [assembly: UserSecretsId("54e8eb06-aaa1-4fff-9f05-3ced1cb623c2")]
 namespace SampleProject.API
